requirements.txt
streamlit
numpy
sklearn
pandas
seaborn
matplotlib
fasttext
tensorflow-cpu